requirements.txt
allure-pytest==2.8.40
allure-python-commons==2.8.40
apipkg==1.5
Appium-Python-Client==1.1.0
astroid==2.5.1
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
execnet==1.8.0
Faker==8.1.2
icdiff==1.9.1
idna==2.10
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
packaging==20.9
pluggy==0.13.1
pprintpp==0.4.0
py==1.10.0
pylint==2.7.2
pyparsing==2.4.7
pytest==6.2.2
pytest-forked==1.3.0
pytest-icdiff==0.5
pytest-xdist==2.2.1
python-dateutil==2.8.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.4
webdriver-manager==3.3.0
wrapt==1.12.1
